[BUGFIX] deleteRecord_flexFormCallBack deletes file_references
Files that are referenced in a flexform
<type>group</type><internal_type>file_reference</internal_type>
will be deleted by deleteRecord_flexFormCallBack() in
t3lib_TCEmain, independent of the path or usage by other records.

For normal record fields (e.g. tt_content.images), only file
with the internal_type "file" are deleted. Files of
internal_type "file_reference" are respected and will not been
deleted. There is an appropriate condition in
t3lib_TCEmain::extFileFunctions().

But for flexform fields, this differentiation in handling
between both internal_types is missing.

Therefore using lowlevel_cleaner or the delete option in the
recycler will delete files that may be still in use. The
flexFormCallBack must be changed to delete only files of
internal_type "file".

Fixes: #40052
Releases: 6.2, 6.1, 6.0, 4.7, 4.5

+++ typo3/sysext/core/Classes/DataHandling/DataHandler.php
@@ -4150,8 +4150,8 @@ class DataHandler {
 		// Use reference index object to find files in fields:
 		$refIndexObj = \TYPO3\CMS\Core\Utility\GeneralUtility::makeInstance('TYPO3\\CMS\\Core\\Database\\ReferenceIndex');
 		$files = $refIndexObj->getRelations_procFiles($dataValue, $dsArr['TCEforms']['config'], $PA['uid']);
-		// Traverse files and delete them:
-		if (is_array($files)) {
+		// Traverse files and delete them if the field is a regular file field (and not a file_reference field)
+		if (is_array($files) && $dsArr['TCEforms']['config']['internal_type'] == 'file') {
 			foreach ($files as $dat) {
 				if (@is_file($dat['ID_absFile'])) {
 					unlink($dat['ID_absFile']);
